♻️(indicators) move daily indicator classes to project core
Daily indicator calculation classes are moved from the dedicated plugin to the
core of the project. These classes are now accessible for reuse across other
plugins as mixins, as they provide generic functionality. Updated all relevant
imports and references to reflect the new structure.
