+++ src/api/core/warren/indicators/mixins.py
@@ -6,14 +6,20 @@ import json
 import logging
 from abc import ABC, abstractmethod
 from functools import cached_property, reduce
-from typing import Any, List, Literal, Protocol, Sequence, Union
+from typing import Any, List, Literal, Optional, Protocol, Sequence, Union
 
 import arrow
+import pandas as pd
 from pydantic.main import BaseModel
+from ralph.backends.data.async_lrs import LRSStatementsQuery
 from sqlmodel import Session, select
 
 from warren.db import get_session as get_db_session
 from warren.filters import DatetimeRange
+from warren.indicators import BaseIndicator
+from warren.models import DailyCount, DailyCounts, DailyUniqueCount, DailyUniqueCounts
+from warren.utils import pipe
+from warren.xapi import StatementsTransformer
 
 from .models import CacheEntry, CacheEntryCreate
 
@@ -271,3 +277,179 @@ class IncrementalCacheMixin(CacheMixin, CacheableIncrementally, ABC):
         values = [self._raw_or_pydantic(cache.value) for cache in caches]
 
         return reduce(self.merge, values)
+
+
+class BaseDailyEvent(BaseIndicator, IncrementalCacheMixin):
+    """Base Daily Event indicator.
+
+    This class defines a base daily event indicator. Given an event type, an object,
+    and a date range, it calculates the total number of events and the number
+    of events per day.
+
+    Required: Indicators inheriting from this base class must declare a 'verb_id'
+    class attribute with their xAPI verb ID.
+    """
+
+    frame: Frames = "day"
+    verb_id: Optional[str] = None
+    object_id: str
+
+    def __init__(
+        self,
+        object_id: str,
+        span_range: DatetimeRange,
+    ):
+        """Instantiate the Daily Event Indicator.
+
+        Args:
+            object_id: The ID of the ressource on which to compute the metric
+            span_range: The date range on which to compute the indicator. It has
+                2 fields, `since` and `until` which are dates or timestamps that must be
+                in ISO format (YYYY-MM-DD, YYYY-MM-DDThh:mm:ss.sss±hh:mm or
+                YYYY-MM-DDThh:mm:ss.sssZ")
+        """
+        super().__init__(span_range=span_range, object_id=object_id)
+
+    def get_lrs_query(
+        self,
+    ) -> LRSStatementsQuery:
+        """Get the LRS query for fetching required statements."""
+        return LRSStatementsQuery(
+            verb=self.verb_id,
+            activity=self.object_id,
+            since=self.since,
+            until=self.until,
+        )
+
+    def filter_statements(self, statements: pd.DataFrame) -> pd.DataFrame:
+        """Filter statements required for this indicator.
+
+        This method may be overridden for indicator-specific filtering.
+        """
+        return statements
+
+    def to_span_range_timezone(self, statements: pd.DataFrame) -> pd.DataFrame:
+        """Convert 'timestamps' column to the DatetimeRange's timezone."""
+        statements = statements.copy()
+        statements["timestamp"] = statements["timestamp"].dt.tz_convert(
+            self.span_range.tzinfo
+        )
+        return statements
+
+    @staticmethod
+    def extract_date_from_timestamp(statements: pd.DataFrame) -> pd.DataFrame:
+        """Convert the 'timestamp' column to a 'date' column with its date values."""
+        statements = statements.copy()
+        statements["date"] = statements["timestamp"].dt.date
+        statements.drop(["timestamp"], axis=1, inplace=True)
+        return statements
+
+
+class DailyEvent(BaseDailyEvent):
+    """Daily Event indicator.
+
+    Required: Indicators inheriting from this base class must declare a 'verb_id'
+    class attribute with their xAPI verb ID.
+    """
+
+    def __init_subclass__(cls, **kwargs):
+        """Ensure subclasses have a 'verb_id' class attribute."""
+        super().__init_subclass__(**kwargs)
+        if cls.verb_id is None:
+            raise TypeError("Indicators must declare a 'verb_id' class attribute")
+
+    async def compute(self) -> DailyCounts:
+        """Fetch statements and computes the current indicator.
+
+        Fetch the statements from the LRS, filter and aggregate them to return the
+        number of activity events per day.
+        """
+        # Initialize daily counts within the specified date range,
+        # with counts equal to zero
+        daily_counts = DailyCounts.from_range(self.since, self.until)
+
+        raw_statements = await self.fetch_statements()
+        if not raw_statements:
+            return daily_counts
+        statements = pipe(
+            StatementsTransformer.preprocess,
+            self.filter_statements,
+            self.to_span_range_timezone,
+            self.extract_date_from_timestamp,
+        )(raw_statements)
+
+        # Compute daily counts from 'statements' DataFrame
+        # and merge them into the 'daily_counts' object
+        daily_counts.merge_counts(
+            [
+                DailyCount(date=date, count=count)
+                for date, count in statements.groupby("date").size().items()
+            ]
+        )
+        return daily_counts
+
+    @staticmethod
+    def merge(a: DailyCounts, b: DailyCounts) -> DailyCounts:
+        """Merging function for computed indicators."""
+        a.merge_counts(b.counts)
+        return a
+
+
+class DailyUniqueEvent(BaseDailyEvent):
+    """Daily Unique Event indicator.
+
+    This class defines a base unique daily event indicator. Given an event
+    type, an activity, and a date range, it calculates the total number of unique
+    user events and the number of unique user events per day.
+    """
+
+    def __init_subclass__(cls, **kwargs):
+        """Ensure subclasses have a 'verb_id' class attribute."""
+        super().__init_subclass__(**kwargs)
+        if cls.verb_id is None:
+            raise TypeError("Indicators must declare a 'verb_id' class attribute")
+
+    def filter_statements(self, statements: pd.DataFrame) -> pd.DataFrame:
+        """Filter statements required for this indicator.
+
+        If necessary, this method removes any duplicate actors from the statements.
+        This filtering step is typically done to ensure that each actor's
+        contributions are counted only once.
+        """
+        return statements.drop_duplicates(subset="actor.uid")
+
+    async def compute(self) -> DailyUniqueCounts:
+        """Fetch statements and computes the current indicator.
+
+        Fetch the statements from the LRS, filter and aggregate them to return the
+        number of unique activity events per day.
+        """
+        # Initialize daily unique counts within the specified date range,
+        # with counts equal to zero
+        daily_unique_counts = DailyUniqueCounts.from_range(self.since, self.until)
+
+        raw_statements = await self.fetch_statements()
+        if not raw_statements:
+            return daily_unique_counts
+
+        statements = pipe(
+            StatementsTransformer.preprocess,
+            self.filter_statements,
+            self.to_span_range_timezone,
+            self.extract_date_from_timestamp,
+        )(raw_statements)
+
+        counts = []
+        for date, users in statements.groupby("date")["actor.uid"].unique().items():
+            counts.append(
+                DailyUniqueCount(date=date, count=len(users), users=set(users))
+            )
+        daily_unique_counts.merge_counts(counts)
+
+        return daily_unique_counts
+
+    @staticmethod
+    def merge(a: DailyUniqueCounts, b: DailyUniqueCounts) -> DailyUniqueCounts:
+        """Merging function for computed indicators."""
+        a.merge_counts(b.counts)
+        return a
