+++ src/api/plugins/video/tests/test_indicators.py
@@ -7,117 +7,23 @@ import urllib
 import httpx
 import pandas as pd
 import pytest
-from dateutil.tz import tzoffset
 from pytest_httpx import HTTPXMock
 from ralph.models.xapi.concepts.constants.video import RESULT_EXTENSION_TIME
 from warren.backends import lrs_client
-from warren.factories.base import BaseXapiStatementFactory
 from warren.filters import DatetimeRange
 from warren.models import DailyUniqueCount
 from warren.xapi import StatementsTransformer
 from warren_video.factories import VideoPlayedFactory
-from warren_video.indicators import BaseDailyEvent, DailyEvent, DailyUniqueViews
-
-
-def test_base_daily_event_subclass_verb_id():
-    """Test '__init_subclass__' for the BaseDailyEvent class."""
-
-    class MyIndicator(DailyEvent):
-        verb_id = "test"
-
-    # the __init_subclass__ is called when the class itself is constructed.
-    # It should raise an error because the 'verb_id' class attribute is missing.
-    with pytest.raises(TypeError) as exception:
-
-        class MyIndicatorMissingAttribute(DailyEvent):
-            wrong_attribute = "test"
-
-    assert str(exception.value) == "Indicators must declare a 'verb_id' class attribute"
-
-
-def test_base_daily_event_span_range_timezone():
-    """Test 'to_span_range_timezone' for the BaseDailyEvent class."""
-    # Create source statements
-    raw_statements = [
-        BaseXapiStatementFactory.build(
-            mutations=[{"timestamp": "2023-01-01T00:10:00.000000+00:00"}]
-        ).dict(),
-        BaseXapiStatementFactory.build(
-            mutations=[{"timestamp": "2023-01-03T00:10:00.000000+00:00"}]
-        ).dict(),
-    ]
-    source_statements = StatementsTransformer.preprocess(raw_statements)
-
-    class MyIndicator(BaseDailyEvent):
-        def merge(self):
-            pass
-
-        def compute(self):
-            pass
-
-    # Create an indicator with a DateTimeRange in UTC
-    indicator_utc = MyIndicator(
-        span_range=DatetimeRange(since="2023-01-01", until="2023-01-03"),
-        video_id="Test",
-    )
-    # Convert statements' timestamp
-    statements = indicator_utc.to_span_range_timezone(source_statements)
-
-    expected_statements = pd.to_datetime(
-        pd.Series(
-            [
-                "2023-01-01T00:10:00+00:00",
-                "2023-01-03T00:10:00+00:00",
-            ],
-            name="timestamp",
-        )
-    )
-
-    # Statements' timestamp should be in UTC.
-    assert statements["timestamp"].equals(expected_statements)
-
-    # Create an indicator with a DateTimeRange in UTC+02:00
-    indicator_with_timezone = MyIndicator(
-        span_range=DatetimeRange(
-            since="2023-01-01T00:00:00+02:00", until="2023-01-03T00:00:00+02:00"
-        ),
-        video_id="Test",
-    )
-    # Convert statements' timestamp
-    statements = indicator_with_timezone.to_span_range_timezone(source_statements)
-
-    # Statements' timestamp should be in UTC+02:00
-    assert statements["timestamp"].equals(
-        expected_statements.dt.tz_convert(tzoffset(None, 7200))
-    )
-
-
-def test_base_daily_event_add_date_column():
-    """Test 'add_date_column' for the BaseDailyEvent class."""
-    raw_statements = [
-        BaseXapiStatementFactory.build(
-            mutations=[{"timestamp": "2023-01-01T00:10:00.000000+00:00"}]
-        ).dict(),
-        BaseXapiStatementFactory.build(
-            mutations=[{"timestamp": "2023-01-03T00:10:00.000000+00:00"}]
-        ).dict(),
-    ]
-    source_statements = StatementsTransformer.preprocess(raw_statements)
-    statements = BaseDailyEvent.extract_date_from_timestamp(source_statements)
-
-    assert statements["date"].equals(
-        pd.to_datetime(pd.Series(["2023-01-01", "2023-01-03"], name="date")).dt.date
-    )
-    assert "timestamp" not in statements.columns
+from warren_video.indicators import DailyUniqueViews
 
 
 @pytest.mark.anyio
 async def test_daily_unique_views(httpx_mock: HTTPXMock, db_session):
     """Test the DailyUniqueViews indicator."""
-    video_id = "uuid://ba4252ce-d042-43b0-92e8-f033f45612ee"
+    object_id = "uuid://ba4252ce-d042-43b0-92e8-f033f45612ee"
 
     local_template = VideoPlayedFactory.template
-    local_template["object"]["id"] = video_id
+    local_template["object"]["id"] = object_id
     local_template["actor"] = {
         "mbox": "mailto:luke@sw.com",
         "objectType": "Agent",
@@ -184,7 +90,7 @@ async def test_daily_unique_views(httpx_mock: HTTPXMock, db_session):
     )
 
     span_range = DatetimeRange(since="2020-01-01", until="2020-01-03")
-    indicator = DailyUniqueViews(video_id=video_id, span_range=span_range)
+    indicator = DailyUniqueViews(object_id=object_id, span_range=span_range)
     daily_counts = await indicator.get_or_compute()
 
     # Generate an example statement to get default actor uid
